+++ crates/tauri-plugin-holochain-service-consumer/build.rs
@@ -1,11 +1,4 @@
-const COMMANDS: &[&str] = &[
-    "connect",
-    "install_app",
-    "enable_app",
-    "is_app_installed",
-    "ensure_app_websocket",
-    "sign_zome_call",
-];
+const COMMANDS: &[&str] = &["setup_app", "enable_app", "sign_zome_call"];
 
 fn main() {
     tauri_plugin::Builder::new(COMMANDS)